HW-7: Apply hilt at RetrofitClient
