+++ class14/simuelunbo/AndroidArchitectureStudy/app/src/main/java/com/example/androidarchitecturestudy/data/api/NaverMovieInterface.kt
@@ -15,38 +15,4 @@ interface NaverMovieInterface {
         @Query("query") query: String,
     ): Call<MovieData>
 
-    companion object {
-        private const val BASE_URL = "https://openapi.naver.com/"
-        private const val CLIENT_ID = "zoFgl2SLHGQXO2WoadKj"
-        private const val CLIENT_SECRET = "Tx2eCHuLZ6"
-
-        fun create(): NaverMovieInterface {
-            val headerInterceptor = Interceptor {
-                val request = it.request()
-                    .newBuilder()
-                    .addHeader(
-                        "X-Naver-Client-Id",
-                        CLIENT_ID
-                    )
-                    .addHeader(
-                        "X-Naver-Client-Secret",
-                        CLIENT_SECRET
-                    )
-                    .build()
-                return@Interceptor it.proceed(request)
-
-            }
-            val client = OkHttpClient.Builder()
-                .addInterceptor(headerInterceptor)
-                .build()
-
-            return Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .client(client)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build()
-                .create(NaverMovieInterface::class.java)
-        }
-
-    }
 }
